requirements.txt
google-genai==1.5.0
chromadb
langchain-community
langchain-text-splitters
pypdf
gradio
python-dotenv